Added crossmatching to find_racs.py
Extensive update to find_racs.py to enable crossmatching to a catalog. Works with an input csv file or coordinates entered via command line. Outputs are the usual fits files and reg/ann/png files if requested and a final crossmatch output csv that includes the components for matched sources. Also added a marker and label to the png files and switched combined images to be default.